+++ ucapi/cover.py
@@ -5,14 +5,14 @@ Cover entity definitions.
 :license: MPL-2.0, see LICENSE for more details.
 """
 
-from enum import Enum
+from enum import StrEnum
 from typing import Any
 
 from .api_definitions import CommandHandler
 from .entity import Entity, EntityTypes
 
 
-class States(str, Enum):
+class States(StrEnum):
     """Cover entity states."""
 
     UNAVAILABLE = "UNAVAILABLE"
@@ -23,7 +23,7 @@ class States(str, Enum):
     CLOSED = "CLOSED"
 
 
-class Features(str, Enum):
+class Features(StrEnum):
     """Cover entity features."""
 
     OPEN = "open"
@@ -35,7 +35,7 @@ class Features(str, Enum):
     TILT_POSITION = "tilt_position"
 
 
-class Attributes(str, Enum):
+class Attributes(StrEnum):
     """Cover entity attributes."""
 
     STATE = "state"
@@ -43,7 +43,7 @@ class Attributes(str, Enum):
     TILT_POSITION = "tilt_position"
 
 
-class Commands(str, Enum):
+class Commands(StrEnum):
     """Cover entity commands."""
 
     OPEN = "open"
@@ -56,7 +56,7 @@ class Commands(str, Enum):
     TILT_STOP = "tilt_stop"
 
 
-class DeviceClasses(str, Enum):
+class DeviceClasses(StrEnum):
     """Cover entity device classes."""
 
     BLIND = "blind"
@@ -68,7 +68,7 @@ class DeviceClasses(str, Enum):
     WINDOW = "window"
 
 
-class Options(str, Enum):
+class Options(StrEnum):
     """Cover entity options."""
 
 
@@ -87,8 +87,11 @@ class Cover(Entity):
         name: str | dict[str, str],
         features: list[Features],
         attributes: dict[str, Any],
+        *,
         device_class: DeviceClasses | None = None,
         options: dict[str, Any] | None = None,
+        icon: str | None = None,
+        description: str | dict[str, str] | None = None,
         area: str | None = None,
         cmd_handler: CommandHandler = None,
     ):
@@ -101,6 +104,8 @@ class Cover(Entity):
         :param attributes: cover attributes
         :param device_class: optional cover device class
         :param options: options
+        :param icon: optional icon
+        :param description: optional description, either a string or a language dictionary
         :param area: optional area
         :param cmd_handler: handler for entity commands
         """
@@ -112,6 +117,8 @@ class Cover(Entity):
             attributes,
             device_class=device_class,
             options=options,
+            icon=icon,
+            description=description,
             area=area,
             cmd_handler=cmd_handler,
         )