+++ ucapi/ir_emitter.py
@@ -5,14 +5,14 @@ IR Emitter entity definitions.
 :license: MPL-2.0, see LICENSE for more details.
 """
 
-from enum import Enum
+from enum import StrEnum
 from typing import Any
 
 from .api_definitions import CommandHandler
 from .entity import Entity, EntityTypes
 
 
-class States(str, Enum):
+class States(StrEnum):
     """IR Emitter entity states."""
 
     UNAVAILABLE = "UNAVAILABLE"
@@ -20,30 +20,30 @@ class States(str, Enum):
     ON = "ON"
 
 
-class Features(str, Enum):
+class Features(StrEnum):
     """IR Emitter entity features."""
 
     SEND_IR = "send_ir"
 
 
-class Attributes(str, Enum):
+class Attributes(StrEnum):
     """IR Emitter entity attributes."""
 
     STATE = "state"
 
 
-class Commands(str, Enum):
+class Commands(StrEnum):
     """IR Emitter entity commands."""
 
     SEND_IR = "send_ir"
     STOP_IR = "stop_ir"
 
 
-class DeviceClasses(str, Enum):
+class DeviceClasses(StrEnum):
     """IR Emitter entity device classes."""
 
 
-class Options(str, Enum):
+class Options(StrEnum):
     """IR Emitter entity options."""
 
     PORTS = "ports"
@@ -67,6 +67,8 @@ class IREmitter(Entity):
         attributes: dict[str, Any],
         *,
         options: dict[str, Any] | None = None,
+        icon: str | None = None,
+        description: str | dict[str, str] | None = None,
         area: str | None = None,
         cmd_handler: CommandHandler = None,
     ):
@@ -78,6 +80,8 @@ class IREmitter(Entity):
         :param features: IR Emitter features
         :param attributes: IR Emitter attributes
         :param options: IR Emitter options
+        :param icon: optional icon
+        :param description: optional description, either a string or a language dictionary
         :param area: optional area
         :param cmd_handler: handler for entity commands
         """
@@ -88,6 +92,8 @@ class IREmitter(Entity):
             features,
             attributes,
             options=options,
+            icon=icon,
+            description=description,
             area=area,
             cmd_handler=cmd_handler,
         )