+++ ucapi/remote.py
@@ -6,7 +6,7 @@ Remote entity definitions.
 """
 
 import dataclasses
-from enum import Enum
+from enum import StrEnum
 from typing import Any
 
 from .api_definitions import CommandHandler
@@ -14,7 +14,7 @@ from .entity import Entity, EntityTypes
 from .ui import DeviceButtonMapping, EntityCommand, UiPage
 
 
-class States(str, Enum):
+class States(StrEnum):
     """Remote entity states."""
 
     UNAVAILABLE = "UNAVAILABLE"
@@ -23,7 +23,7 @@ class States(str, Enum):
     OFF = "OFF"
 
 
-class Features(str, Enum):
+class Features(StrEnum):
     """Remote entity features."""
 
     ON_OFF = "on_off"
@@ -31,13 +31,13 @@ class Features(str, Enum):
     SEND_CMD = "send_cmd"
 
 
-class Attributes(str, Enum):
+class Attributes(StrEnum):
     """Remote entity attributes."""
 
     STATE = "state"
 
 
-class Commands(str, Enum):
+class Commands(StrEnum):
     """Remote entity commands."""
 
     ON = "on"
@@ -47,7 +47,7 @@ class Commands(str, Enum):
     SEND_CMD_SEQUENCE = "send_cmd_sequence"
 
 
-class Options(str, Enum):
+class Options(StrEnum):
     """Remote entity options."""
 
     SIMPLE_COMMANDS = "simple_commands"
@@ -128,9 +128,12 @@ class Remote(Entity):
         name: str | dict[str, str],
         features: list[Features],
         attributes: dict[str, Any],
+        *,
         simple_commands: list[str] | None = None,
         button_mapping: list[DeviceButtonMapping | dict[str, Any]] | None = None,
         ui_pages: list[UiPage | dict[str, Any]] | None = None,
+        icon: str | None = None,
+        description: str | dict[str, str] | None = None,
         area: str | None = None,
         cmd_handler: CommandHandler = None,
     ):
@@ -146,6 +149,8 @@ class Remote(Entity):
                Either with DeviceButtonMapping items or plain dictionary items.
         :param ui_pages: optional user interface page definitions.
                Either with UiPage items or plain dictionary items.
+        :param icon: optional icon
+        :param description: optional description, either a string or a language dictionary
         :param area: optional area
         :param cmd_handler: handler for entity commands
         """
@@ -163,6 +168,8 @@ class Remote(Entity):
             features,
             attributes,
             options=options,
+            icon=icon,
+            description=description,
             area=area,
             cmd_handler=cmd_handler,
         )