+++ ucapi/select.py
@@ -5,14 +5,14 @@ Switch entity definitions.
 :license: MPL-2.0, see LICENSE for more details.
 """
 
-from enum import Enum
+from enum import StrEnum
 from typing import Any
 
 from .api_definitions import CommandHandler
 from .entity import Entity, EntityTypes
 
 
-class States(str, Enum):
+class States(StrEnum):
     """Select entity states."""
 
     UNAVAILABLE = "UNAVAILABLE"
@@ -20,11 +20,11 @@ class States(str, Enum):
     ON = "ON"
 
 
-class Features(str, Enum):
+class Features(StrEnum):
     """Select entity features."""
 
 
-class Attributes(str, Enum):
+class Attributes(StrEnum):
     """Select entity attributes."""
 
     STATE = "state"
@@ -32,7 +32,7 @@ class Attributes(str, Enum):
     OPTIONS = "options"
 
 
-class Commands(str, Enum):
+class Commands(StrEnum):
     """Select entity commands."""
 
     SELECT_OPTION = "select_option"
@@ -42,11 +42,11 @@ class Commands(str, Enum):
     SELECT_PREVIOUS = "select_previous"
 
 
-class DeviceClasses(str, Enum):
+class DeviceClasses(StrEnum):
     """Select entity device classes."""
 
 
-class Options(str, Enum):
+class Options(StrEnum):
     """Select entity options."""
 
 
@@ -65,6 +65,8 @@ class Select(Entity):
         name: str | dict[str, str],
         attributes: dict[str, Any],
         *,
+        icon: str | None = None,
+        description: str | dict[str, str] | None = None,
         area: str | None = None,
         cmd_handler: CommandHandler = None,
     ):
@@ -74,6 +76,8 @@ class Select(Entity):
         :param identifier: entity identifier
         :param name: friendly name
         :param attributes: select attributes
+        :param icon: optional icon
+        :param description: optional description, either a string or a language dictionary
         :param area: optional area
         :param cmd_handler: handler for entity commands
         """
@@ -83,6 +87,8 @@ class Select(Entity):
             EntityTypes.SELECT,
             [],
             attributes,
+            icon=icon,
+            description=description,
             area=area,
             cmd_handler=cmd_handler,
         )