+++ ucapi/switch.py
@@ -5,14 +5,14 @@ Switch entity definitions.
 :license: MPL-2.0, see LICENSE for more details.
 """
 
-from enum import Enum
+from enum import StrEnum
 from typing import Any
 
 from .api_definitions import CommandHandler
 from .entity import Entity, EntityTypes
 
 
-class States(str, Enum):
+class States(StrEnum):
     """Switch entity states."""
 
     UNAVAILABLE = "UNAVAILABLE"
@@ -21,20 +21,20 @@ class States(str, Enum):
     OFF = "OFF"
 
 
-class Features(str, Enum):
+class Features(StrEnum):
     """Switch entity features."""
 
     ON_OFF = "on_off"
     TOGGLE = "toggle"
 
 
-class Attributes(str, Enum):
+class Attributes(StrEnum):
     """Switch entity attributes."""
 
     STATE = "state"
 
 
-class Commands(str, Enum):
+class Commands(StrEnum):
     """Switch entity commands."""
 
     ON = "on"
@@ -42,14 +42,14 @@ class Commands(str, Enum):
     TOGGLE = "toggle"
 
 
-class DeviceClasses(str, Enum):
+class DeviceClasses(StrEnum):
     """Switch entity device classes."""
 
     OUTLET = "outlet"
     SWITCH = "switch"
 
 
-class Options(str, Enum):
+class Options(StrEnum):
     """Switch entity options."""
 
     READABLE = "readable"
@@ -70,8 +70,11 @@ class Switch(Entity):
         name: str | dict[str, str],
         features: list[Features],
         attributes: dict[str, Any],
+        *,
         device_class: DeviceClasses | None = None,
         options: dict[str, Any] | None = None,
+        icon: str | None = None,
+        description: str | dict[str, str] | None = None,
         area: str | None = None,
         cmd_handler: CommandHandler = None,
     ):
@@ -84,6 +87,8 @@ class Switch(Entity):
         :param attributes: switch attributes
         :param device_class: optional switch device class
         :param options: options
+        :param icon: optional icon
+        :param description: optional description, either a string or a language dictionary
         :param area: optional area
         :param cmd_handler: handler for entity commands
         """
@@ -95,6 +100,8 @@ class Switch(Entity):
             attributes,
             device_class=device_class,
             options=options,
+            icon=icon,
+            description=description,
             area=area,
             cmd_handler=cmd_handler,
         )